src: Re-use mask texture; Respect linesize; Enable Accurate mode
For Accurate, the image width is 2016 but the linesize is 2048, which
lead to everything shifting. gs_texture_set_image lets us specify a
different linesize than what would be "expected" by the color format. We
can also re-use the mask texture, which is more performant than
destroying and recreating each time and probably should have always been
done that way.

Co-authored-by: thebabush <1985669+thebabush@users.noreply.github.com>